fix: 자료형 변경

+++ Vinyla/Vinyla/Modules/AddReview/AddReviewViewModel.swift
@@ -16,7 +16,7 @@ protocol AddReviewViewModelProtocol {
 }
 
 final class AddReviewViewModel: AddReviewViewModelProtocol {
-    var songRate: Int?
+    var songRate: Double?
     var songRateCount: Int?
     //input
     var userRate: PublishSubject<Int> = PublishSubject<Int>()
